+++ docs/changelog.rst
@@ -264,6 +264,9 @@ Detailed list of changes
 
 - Cursor trail: Show a cursor trail when switching tabs (:pull:`9588`)
 
+- Make shift+left click extend the current selection instead of starting a new
+  selection when the mouse is not grabbed by the TUI application (:disc:`9608`)
+
 
 0.45.0 [2025-12-24]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~